wif_FastAPI/pyproject.toml
[tool.poetry]
name = "wif-fastapi"
version = "0.1.0"
description = ""
authors = ["Hordunlarmy <Hordunlarmy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
sqlalchemy = "^2.0.40"
python-decouple = "^3.8"
starlette-wtf = "^0.4.5"
python-jose = "^3.4.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pycryptodome = "^3.22.0"
jinja2 = "^3.1.6"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
